Rename inventory folder to routes, separate routes by action type
