--- src/inventory/index.js
@@ -1,8 +1,0 @@
-import buy from './buy'
-import sell from './sell'
-import queue from './queue'
-
-export default [
-  buy
-  ,sell
-]
