+++ src/routes/index.js
@@ -1,0 +1,9 @@
+import buy from './buy'
+import sell from './sell'
+
+export default {
+  POST: [
+    buy
+    ,sell
+  ]
+}
